Move esbuild plugins to their own subfolder
