+++ scripts/esbuild-plugins/flow-remove-types-plugin.js
