+++ scripts/esbuild-plugins/vue-inline-svg-plugin.js
