+++ scripts/esbuild-plugins/vue-plugin.js
